requirements.txt
tweepy
kafka-python
elasticsearch==8.12.0
python-dotenv
pyspark